Numpy - 2

+++ OOPS/numpy_tutorial.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-a = np.array([1,2,3,4,5])
+a = np.array([1,2,3,4,5], dtype="int32")
 b = np.array([[1.0,2.0,3.0], [4.0,5.0, 6.0]])
 #print(b.ndim)
-
+#print(b.shape, a.itemsize)
